Add L10NSharp localization manager
Also rename the project
