+++ .idea/.idea.USFMSharp/.idea/markdown-navigator/profiles_settings.xml
@@ -1,0 +1,3 @@
+<component name="MarkdownNavigator.ProfileManager">
+  <settings default="" />
+</component>
