+++ .idea/.idea.USFMSharp/riderModule.iml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module type="RIDER_MODULE" version="4">
+  <component name="NewModuleRootManager">
+    <content url="file://$MODULE_DIR$/../..">
+      <sourceFolder url="file://$MODULE_DIR$/../../USFMSharp" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/../../USFMSharpTests" isTestSource="false" />
+      <excludeFolder url="file://$MODULE_DIR$/../../USFMSharp/bin" />
+      <excludeFolder url="file://$MODULE_DIR$/../../USFMSharp/obj" />
+      <excludeFolder url="file://$MODULE_DIR$/../../USFMSharpTests/bin" />
+      <excludeFolder url="file://$MODULE_DIR$/../../USFMSharpTests/obj" />
+      <excludeFolder url="file://$MODULE_DIR$/../../packages" />
+    </content>
+    <orderEntry type="sourceFolder" forTests="false" />
+  </component>
+</module>
