+++ README.md
@@ -1,4 +1,4 @@
-# uw-edit: a cross-platform, standards-compliant USFM editor
+# USFMSharp: a cross-platform, standards-compliant USFM editor
 
 The goal of this project is to produce a USFM editor that will allow anyone to
 create and edit USFM files that are 100% in compliance with the USFM standards
