+++ USFMSharp.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "USFMSharpTests", "USFMSharpTests\USFMSharpTests.csproj", "{B52F7B64-9561-4C9D-AC72-942A37A9392E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "USFMSharp", "USFMSharp\USFMSharp.csproj", "{188B2703-F0C4-486A-AAF8-1F87C9EAECE6}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B52F7B64-9561-4C9D-AC72-942A37A9392E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B52F7B64-9561-4C9D-AC72-942A37A9392E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B52F7B64-9561-4C9D-AC72-942A37A9392E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B52F7B64-9561-4C9D-AC72-942A37A9392E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{188B2703-F0C4-486A-AAF8-1F87C9EAECE6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{188B2703-F0C4-486A-AAF8-1F87C9EAECE6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{188B2703-F0C4-486A-AAF8-1F87C9EAECE6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{188B2703-F0C4-486A-AAF8-1F87C9EAECE6}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(MonoDevelopProperties) = preSolution
+		version = 0.0.0.1
+	EndGlobalSection
+EndGlobal
