+++ USFMSharp/SIL/PlatformUtilities/Platform.cs
