+++ USFMSharp/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="L10NSharp" version="2.0.51" targetFramework="net45" />
+</packages>
