--- uw-edit-tests/Views/MainViewModelTests.cs
@@ -1,17 +1,0 @@
-﻿using NUnit.Framework;
-using uw_edit.Views;
-
-namespace uw_edit_tests
-{
-	[TestFixture]
-	public class MainViewModelTests
-	{
-		[Test]
-		public void CreateInstance()
-		{
-			var mainViewModel = new MainViewModel();
-			Assert.IsInstanceOf<MainViewModel>(mainViewModel);
-		}
-	}
-}
-
