--- uw-edit/packages.Linux
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Geckofx45.32.Linux" version="45.0.21.0" targetFramework="net45" />
-  <package id="Geckofx45.64.Linux" version="45.0.21.0" targetFramework="net45" />
-  <package id="L10NSharp" version="2.0.51" targetFramework="net452" />
-</packages>
