--- uw-edit/packages.Windows
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Geckofx45" version="45.0.20" targetFramework="net452" />
-  <package id="L10NSharp" version="2.0.51" targetFramework="net452" />
-</packages>
