--- uw-edit/uw-edit.csproj
@@ -1,86 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{188B2703-F0C4-486A-AAF8-1F87C9EAECE6}</ProjectGuid>
-    <OutputType>WinExe</OutputType>
-    <RootNamespace>uw_edit</RootNamespace>
-    <AssemblyName>uw-edit</AssemblyName>
-    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
-    <NuGetPackageImportStamp>
-    </NuGetPackageImportStamp>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug</OutputPath>
-    <DefineConstants>DEBUG;</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <ConsolePause>false</ConsolePause>
-    <CheckForOverflowUnderflow>true</CheckForOverflowUnderflow>
-    <PlatformTarget>x86</PlatformTarget>
-    <Commandlineparameters>/home/team43/Projects/uw-edit/test-data/41-MAT-ULB.usfm</Commandlineparameters>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release</OutputPath>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <ConsolePause>false</ConsolePause>
-    <PlatformTarget>x86</PlatformTarget>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="Views\MainView.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="Views\MainViewModel.cs">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Compile>
-    <Compile Include="UserControls\MainViewMenu.cs">
-      <SubType>Component</SubType>
-    </Compile>
-    <Compile Include="Program.cs" />
-    <Compile Include="UserControls\MainViewStrip.cs">
-      <SubType>Component</SubType>
-    </Compile>
-    <Compile Include="SIL\PlatformUtilities\Platform.cs" />
-    <Compile Include="USFM\TextTools.cs" />
-    <Compile Include="USFM\ParatextStylesheet.cs" />
-    <Compile Include="Tools\Diff.cs" />
-    <Compile Include="UserControls\MainViewStatusStrip.cs" />
-    <Compile Include="Tools\UnixTimestamp.cs" />
-    <Compile Include="USFM\TagError.cs" />
-    <Compile Include="Views\SaveChangesDialog.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <Reference Include="System.Windows.Forms" />
-    <Reference Include="System" />
-    <Reference Include="System.Drawing" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="App.config" />
-    <None Include="packages.config" />
-    <None Include="Resources\usfm_sb.sty">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Include="Resources\usfm_template.usfm">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Include="Resources\save.png">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-  <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
-  <Target Name="SelectTargetsConfig">
-    <Exec Condition="'$(OS)'=='Windows_NT'" Command="copy /Y /B packages.Windows packages.config" />
-    <Exec Condition="'$(OS)'!='Windows_NT'" Command="cat packages.Linux &gt;packages.config" />
-  </Target>
-  <ItemGroup />
-  <ItemGroup>
-    <Folder Include="Resources\" />
-    <Folder Include="Tools\" />
-  </ItemGroup>
-</Project>
